feat: agregar funcionalidad de la barra de progreso d electura, guardar el porcentaje de progreso y mostrarlo en cada libro, agregar funcionalidad al botón de atrás en el área de lectura y agregar la vista del menú de opciones (preferencias de usuario + más opciones)

+++ app/src/main/java/tfg/carlos/wereaderapp/data/entity/BookEntity.kt
@@ -28,7 +28,7 @@ data class BookEntity(
     // Campos locales ROOM:
     val isPending: Boolean = false,
     val isReading: Boolean = false,
-    val readingProgress: Int = 0, // Porcentaje de lectura (0-100)
+    val readingProgress: Double = 0.0, // Porcentaje de lectura (0-100)
     val mine: Boolean = true,
     val idUser: String,
     val lastLocator: String? = null, // Última posición de lectura (JSON de Readium) (null si no se ha leído el libro aún)
